delete, add operations for MedicationEntity.

+++ src/main/java/kz/medicare/repository/MedicationRemindRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface MedicationRemindRepository extends JpaRepository<MedicationRemind, Long> {
+public interface MedicationRemindRepository extends JpaRepository<MedicationRemind, Integer> {
 
     List<MedicationRemind> findMedicationRemindsByPatient(@NotNull User patient);
 }
